Adopt typedoc for API docs (#66), closes #63
* Build the API docs when the site is being built

* Add watch:docs command

* Add a custom CSS stub file

* Use a custom index page for the API docs

* Add support for the abstract tag

* Hide the protected and external filters since we don't have such modifiers

* Update inline docs so that Typedoc parses them correctly

* Add (partial) support for Prism

In Typedoc, we can use no extensibility point [1] to replace the built-in syntax highlighter (Shiki) with Prism. The only way to do that is to write a custom theme that (somehow) uses another syntax highlighter or doesn't highlight code blocks at all. I tried but failed — there's a ton of stuff to know to be able to do that (at least from my perspective).

However, we can make the built-in markdown plugin (Marked) use a custom highlight function instead of the one that calls the Typedoc built-in syntax highlighter. That means that our markdown files and code blocks in the code comments (Typedoc supports markdown in comments) can be highlighted by Prism.

[1] https://discord.com/channels/508357248330760243/829307039447515176/1202594912051273848